+++ tests/ui/c/src/t_ui_framerate.c
@@ -1,0 +1,71 @@
+/*
+ * C
+ *
+ * Copyright 2019 MicroEJ Corp. All rights reserved.
+ * For demonstration purpose only.
+ * MicroEJ Corp. PROPRIETARY/CONFIDENTIAL. Use is subject to license terms.
+ */
+#include <stdio.h>
+#include <stdarg.h>
+#include "../../../../framework/c/embunit/embUnit/embUnit.h"
+#include "u_print.h"
+#include "x_version.h"
+#include "t_ui_framerate.h"
+#include "x_ui_framerate.h"
+#include "LLDISPLAY.h"
+#include "u_time_base.h"
+
+static void T_UI_FRAMERATE_setUp(void)
+{
+	UTIL_print_initialize();
+	LLDISPLAY_initialize();
+	UTIL_TIME_BASE_initialize();
+}
+
+static void T_UI_FRAMERATE_tearDown(void)
+{
+
+}
+
+static void print_drawing_time_report(float frequency, uint32_t div, uint32_t framerate_time_us, uint32_t flush_copy_time_us)
+{
+	UTIL_print_string("To have an animation at ");
+	UTIL_print_float(frequency / div);
+	UTIL_print_string(" Hz, the drawing time cannot be higher than ");
+	UTIL_print_float(((float)(framerate_time_us * div - flush_copy_time_us))/1000);
+	UTIL_print_string(" ms.\n");
+}
+
+static void T_UI_FRAMERATE_getFramerate(void)
+{
+	uint32_t buffer_address = LLDISPLAY_getGraphicsBufferAddress();
+	uint32_t framerate_time_us;
+	uint32_t flush_copy_time_us;
+
+	X_UI_FRAMERATE_get(LLDISPLAY_getGraphicsBufferAddress(), &framerate_time_us, &flush_copy_time_us);
+
+	float frequency = 1/(((float)framerate_time_us)/1000000);
+
+        UTIL_print_string("\n");
+	UTIL_print_string("Flush copy time is ");
+	UTIL_print_float(((float)framerate_time_us)/1000);
+	UTIL_print_string(" ms (");
+	UTIL_print_float(frequency);
+	UTIL_print_string(" Hz)\n");
+
+	print_drawing_time_report(frequency, 1, framerate_time_us, flush_copy_time_us);
+	print_drawing_time_report(frequency, 2, framerate_time_us, flush_copy_time_us);
+	print_drawing_time_report(frequency, 3, framerate_time_us, flush_copy_time_us);
+}
+
+TestRef T_UI_FRAMERATE_tests(void)
+{
+	EMB_UNIT_TESTFIXTURES(fixtures)
+	{
+		new_TestFixture("Framerate", T_UI_FRAMERATE_getFramerate),
+	};
+
+	EMB_UNIT_TESTCALLER(framerateTest, "Framerate_tests", T_UI_FRAMERATE_setUp, T_UI_FRAMERATE_tearDown, fixtures);
+
+	return (TestRef)&framerateTest;
+}