+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-  - Add UI test `t_ui_api.c`.
+  - Add UI test `t_ui_api.c`: test LLDisplay implementation.
+  - Add UI test `t_ui_framerate.c`: retrieve LCD framerate.
   
 ## 1.0.1 - 2019-04-25